requirements.txt
asgiref==3.5.2
certifi==2022.12.7
dj-database-url==1.2.0
Django==4.1
django-extensions==3.2.1
gunicorn==20.1.0
pip==23.0.1
psycopg2-binary==2.9.5
python-decouple==3.8
python-dotenv==1.0.0
setuptools==65.6.3
sqlparse==0.4.3
wheel==0.38.4
whitenoise==6.4.0
